Read Fed Files

Read Fed Files

Usage

read_fed(filename, lib_version = NULL, deduplicate_method = "offset")

Arguments

filename

file path of csv file produced by FED3

lib_version

character with library version to be used for specifying the columns

deduplicate_method

Method to be implemented by deduplicate_datetime() to deal with duplicate timestamps recorded by FED machines.

Value

A data.frame with parsed columns added to the original FED data.

Details

This function will read the raw data with column specifications (see fed_col_types()). It will append the datetime by formatting the FED's clock datetime as yyyy-mm-dd HH:MM:SS. It will parse the session from the filename and append it to the resulting data.frame.

See also